+++ OggExport/src/stream_audio.c
@@ -668,7 +668,7 @@ ComponentResult write_i_header__audio(StreamInfoPtr si, DataHandler data_h,
                 header.e_o_s = 0;
                 header.granulepos = 0;
                 header.packetno = 0;
-                header.bytes = EndianS32_BtoN(aheader->size) - 2 * sizeof(long);
+                header.bytes = EndianS32_BtoN(aheader->size) - 2 * sizeof(int);
                 header.packet = aheader->data;
                 break;
 
@@ -677,7 +677,7 @@ ComponentResult write_i_header__audio(StreamInfoPtr si, DataHandler data_h,
                 header_vc.e_o_s = 0;
                 header_vc.granulepos = 0;
                 header_vc.packetno = 1;
-                header_vc.bytes = EndianS32_BtoN(aheader->size) - 2 * sizeof(long);
+                header_vc.bytes = EndianS32_BtoN(aheader->size) - 2 * sizeof(int);
                 header_vc.packet = aheader->data;
                 break;
 
@@ -686,7 +686,7 @@ ComponentResult write_i_header__audio(StreamInfoPtr si, DataHandler data_h,
                 header_cb.e_o_s = 0;
                 header_cb.granulepos = 0;
                 header_cb.packetno = 2;
-                header_cb.bytes = EndianS32_BtoN(aheader->size) - 2 * sizeof(long);
+                header_cb.bytes = EndianS32_BtoN(aheader->size) - 2 * sizeof(int);
                 header_cb.packet = aheader->data;
                 break;
