+++ OggExport/src/stream_video.c
@@ -898,7 +898,7 @@ ComponentResult write_i_header__video(StreamInfoPtr si, DataHandler data_h,
                 header.e_o_s = 0;
                 header.granulepos = 0;
                 header.packetno = 0;
-                header.bytes = EndianS32_BtoN(aheader->size) - 2 * sizeof(long);
+                header.bytes = EndianS32_BtoN(aheader->size) - 2 * sizeof(int);
                 header.packet = aheader->data;
                 break;
 
@@ -907,7 +907,7 @@ ComponentResult write_i_header__video(StreamInfoPtr si, DataHandler data_h,
                 header_tc.e_o_s = 0;
                 header_tc.granulepos = 0;
                 header_tc.packetno = 1;
-                header_tc.bytes = EndianS32_BtoN(aheader->size) - 2 * sizeof(long);
+                header_tc.bytes = EndianS32_BtoN(aheader->size) - 2 * sizeof(int);
                 header_tc.packet = aheader->data;
                 break;
 
@@ -916,7 +916,7 @@ ComponentResult write_i_header__video(StreamInfoPtr si, DataHandler data_h,
                 header_cb.e_o_s = 0;
                 header_cb.granulepos = 0;
                 header_cb.packetno = 2;
-                header_cb.bytes = EndianS32_BtoN(aheader->size) - 2 * sizeof(long);
+                header_cb.bytes = EndianS32_BtoN(aheader->size) - 2 * sizeof(int);
                 header_cb.packet = aheader->data;
                 break;
